+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - [RT#187856]: repeated reminders fail.
+- [353](https://github.com/CCSDForge/episciences/issues/353): fix doi queue not saved + fix delete volume (alter the table for volume doi queue).
 ## 1.0.40.7 - 2023-06-08
 ### Fixed
 - [133](https://github.com/CCSDForge/episciences/issues/133): wrong acceptance date and docUrl for tmp versions.